feat(report): process SQS messages to update total and count of approved loans

+++ domain/model/src/main/java/co/com/pragma/model/approvedloans/ApprovedLoans.java
@@ -13,6 +13,6 @@ import lombok.Setter;
 public class ApprovedLoans {
     private String id;
     private Integer totalCount;
-    private Integer totalAmount;
+    private Double totalAmount;
     private String lastUpdated;
 }
