+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/pragma/api/dto/ApprovedLoansDTO.java
@@ -3,7 +3,7 @@ package co.com.pragma.api.dto;
 public record ApprovedLoansDTO(
     String id,
     Integer totalCount,
-    Long totalAmount,
+    Double totalAmount,
     String lastUpdated
 ) {
 }
